+++ app/Http/Controllers/SubbredditsController.php
@@ -44,9 +44,12 @@ class SubbredditsController extends Controller
      */
     public function show($id)
     {
-        return \App\Subbreddit::with(['posts.comments.childComments','user'])->find($id);
+        return \App\Subbreddit::with([
+            'posts.comments.childComments',
+            'user'
+        ])->find($id);
     }
-    
+
     /**
      * Update the specified resource in storage.
      *
@@ -57,9 +60,13 @@ class SubbredditsController extends Controller
     public function update(Request $request, $id)
     {
         $subbreddit = \App\Subbreddit::find($id);
-        $subbreddit->name = $request->name;
-        $subbreddit->description = $request->description;
-        $subbreddit->save();
+        if ($subbreddit->user_id == \Auth::user()->id) {
+            $subbreddit->name = $request->name;
+            $subbreddit->description = $request->description;
+            $subbreddit->save();
+        } else {
+            return response("Unauthorized", 403);
+        }
 
         return $subbreddit;
     }
@@ -72,9 +79,12 @@ class SubbredditsController extends Controller
      */
     public function destroy($id)
     {
-         $subbreddit = \App\Subbreddit::find($id);
-         $subbreddit->delete();
-         return $subbreddit;
-
+        $subbreddit = \App\Subbreddit::find($id);
+        if ($subbreddit->user_id == \Auth::user()->id) {
+            $subbreddit->delete();
+        } else {
+            return response("Unauthorized", 403);
+        }
+        return $subbreddit;
     }
 }
